Move image models to gammapy.modeling
