+++ gammapy/catalog/fermi.py
@@ -6,8 +6,8 @@ import astropy.units as u
 from astropy.table import Column, Table
 from astropy.time import Time
 from gammapy.cube.models import SkyModel
-from gammapy.image.models import SkyDiffuseMap, SkyDisk, SkyGaussian, SkyPointSource
 from gammapy.maps import Map
+from gammapy.modeling.models import SkyDiffuseMap, SkyDisk, SkyGaussian, SkyPointSource
 from gammapy.spectrum import FluxPoints
 from gammapy.spectrum.models import (
     ExponentialCutoffPowerLaw3FGL,
