+++ gammapy/cube/tests/test_fit.py
@@ -7,13 +7,13 @@ from astropy.coordinates import SkyCoord
 from regions import CircleSkyRegion
 from gammapy.cube import MapDataset, PSFKernel, make_map_exposure_true_energy
 from gammapy.cube.models import BackgroundModel, SkyModel
-from gammapy.image.models import SkyGaussian
 from gammapy.irf import (
     EffectiveAreaTable2D,
     EnergyDependentMultiGaussPSF,
     EnergyDispersion,
 )
 from gammapy.maps import Map, MapAxis, WcsGeom
+from gammapy.modeling.models import SkyGaussian
 from gammapy.spectrum.models import PowerLaw
 from gammapy.utils.fitting import Fit
 from gammapy.utils.testing import mpl_plot_check, requires_data, requires_dependency
