+++ gammapy/cube/tests/test_models.py
@@ -11,9 +11,9 @@ from gammapy.cube.models import (
     SkyModel,
     SkyModels,
 )
-from gammapy.image.models import SkyGaussian
 from gammapy.irf import EnergyDispersion
 from gammapy.maps import Map, MapAxis, WcsGeom
+from gammapy.modeling.models import SkyGaussian
 from gammapy.spectrum.models import PowerLaw
 from gammapy.utils.testing import requires_data
