+++ gammapy/cube/tests/test_simulate.py
@@ -5,9 +5,9 @@ import astropy.units as u
 from astropy.coordinates import SkyCoord
 from gammapy.cube import MapDataset, simulate_dataset
 from gammapy.cube.models import SkyModel, SkyModels
-from gammapy.image.models import SkyGaussian
 from gammapy.irf import load_cta_irfs
 from gammapy.maps import MapAxis, WcsGeom
+from gammapy.modeling.models import SkyGaussian
 from gammapy.spectrum.models import PowerLaw
 from gammapy.utils.testing import requires_data
