--- gammapy/image/__init__.py
@@ -1,2 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Sky images (2-dimensional: lon, lat)."""
