+++ gammapy/modeling/models/__init__.py
@@ -1,3 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Built-in models in Gammapy."""
+from .image import *
 from .time import *
