+++ gammapy/modeling/models/image/__init__.py
