+++ gammapy/modeling/models/image/core.py
