+++ gammapy/utils/random/tests/test_inverse_cdf.py
@@ -8,10 +8,9 @@ from astropy.table import Table
 from astropy.time import Time
 from gammapy.cube import MapEvaluator
 from gammapy.cube.models import SkyModel
-from gammapy.image.models import SkyGaussian
 from gammapy.maps import Map, MapAxis
+from gammapy.modeling.models import LightCurveTableModel, SkyGaussian
 from gammapy.spectrum.models import PowerLaw
-from gammapy.time.models import LightCurveTableModel
 from gammapy.utils.random import InverseCDFSampler, MapEventSampler
