+++ gammapy/utils/serialization/tests/test_io.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 from astropy.utils.data import get_pkg_data_filename
 from gammapy.cube.models import BackgroundModels, SkyModels
-from gammapy.image import models as spatial
+from gammapy.modeling.models import image as spatial
 from gammapy.spectrum import models as spectral
 from gammapy.utils.fitting import Datasets
 from gammapy.utils.scripts import read_yaml
